Moves IdentityHash out of lib.rs
Simplifies the lib.rs file

Inlines the const_assert macro that is only used once
